handlers: split processor handler funcs into modules
handlers: split program_tests into multiple modules

handlers: added convenient make targest for running specific test modules
