+++ src/handlers.rs
@@ -1,5 +1,13 @@
 pub mod account_settings_update_handler;
+pub mod approval_disposition_handler;
+pub mod balance_account_creation_handler;
+pub mod balance_account_update_handler;
 pub mod dapp_book_update_handler;
 pub mod dapp_transaction_handler;
+pub mod init_wallet_handler;
+pub mod transfer_handler;
+pub mod update_signer_handler;
 pub mod utils;
 pub mod wallet_config_policy_update_handler;
+pub mod wallet_update_handler;
+pub mod wrap_unwrap_handler;
